requirements.txt
# Core
python-telegram-bot==20.7
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
asyncio==3.4.3
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
sqlite3-utils==3.35
chromadb==0.4.18
redis==5.0.1

# LLM APIs
openai==1.6.0
google-generativeai==0.3.2
anthropic==0.8.1
httpx==0.25.2

# Vector & Embeddings
sentence-transformers==2.2.2
numpy==1.26.2

# Utils
pyyaml==6.0.1
schedule==1.2.1
python-dateutil==2.8.2
emoji==2.9.0
regex==2023.10.3

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
